requirements.txt
requests
confluent-kafka
faker
pyspark
cassandra-driver
jsonschema
elasticsearch
